Add better Error Handlers & fix some Sintax Errors

+++ controllers/adminController.js
@@ -1,3 +1,4 @@
+import adminHandler from "../utils/errorHandler.js";
 import Admin from "../models/adminModel.js";
 import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
@@ -5,17 +6,27 @@ import "dotenv/config";
 
 async function createAdmin(req, res) {
   try {
+    const existingAdmin = await Admin.findOne({ email: req.body.email });
+    if (existingAdmin) {
+      adminHandler.handleDuplicateError(res, "The email is already registered");
+      return;
+    }
+
     const newAdmin = await Admin.create({
       adminName: req.body.adminName,
       name: req.body.name,
       surname: req.body.surname,
       email: req.body.email,
       password: req.body.password,
-      profileImage: req.file.filename
+      profileImage: req.file.filename,
     });
     res.json(newAdmin);
   } catch (error) {
-    res.status(500).json("The server had an error");
+    if (error.name === "ValidationError") {
+      adminHandler.handleValidationError(res, error.message);
+    } else {
+      adminHandler.handleServerError(res);
+    }
   }
 }
 
@@ -23,6 +34,11 @@ async function editAdmin(req, res) {
   try {
     const { id } = await Admin.findById(req.auth.sub);
     const foundAdmin = await Admin.findById(req.params.id);
+
+    if (!foundAdmin) {
+      adminHandler.handleNotFoundError(res, "Admin");
+      return;
+    }
 
     if (id === foundAdmin.id) {
       foundAdmin.adminName = req.file.adminName ?? foundAdmin.adminName;
@@ -36,10 +52,14 @@ async function editAdmin(req, res) {
 
       res.json(foundAdmin);
     } else {
-      res.json("You cannot edit this Admin, check again");
+      adminHandler.handleAuthError(res);
     }
   } catch (error) {
-    res.status(500).json("The server had an error");
+    if (error.name === "ValidationError") {
+      adminHandler.handleValidationError(res, error.message);
+    } else {
+      adminHandler.handleServerError(res);
+    }
   }
 }
 
@@ -48,37 +68,45 @@ async function deleteAdmin(req, res) {
     const { id } = await Admin.findById(req.auth.sub);
     const foundAdmin = await Admin.findById(req.params.id);
 
+    if (!foundAdmin) {
+      adminHandler.handleNotFoundError(res, "Admin");
+      return;
+    }
+
     if (id === foundAdmin.id) {
-      const deletedAdmin = await Admin.findByIdAndDelete(req.params.id);
+      await Admin.findByIdAndDelete(req.params.id);
       res.json("The Admin was deleted");
     } else {
-      res.json("You cannot delete this Admin, check again");
+      adminHandler.handleAuthError(res);
     }
   } catch (error) {
-    res.status(500).json("The server had an error");
+    adminHandler.handleServerError(res);
   }
 }
 
 async function loginAdmin(req, res) {
   try {
     const admin = await Admin.findOne({ email: req.body.email });
-    if (admin !== null) {
-      const validHash = await bcrypt.compare(req.body.password, admin.password);
-      if (validHash) {
-        const tokenPayLoad = {
-          sub: admin.id,
-          iat: Date.now(),
-        };
-        const token = jwt.sign(tokenPayLoad, process.env.JWT_KEY);
-        res.json({ token: token });
-      } else {
-        res.json("invalid credentials");
-      }
+
+    if (!admin) {
+      handleAuthError(res);
+      return;
+    }
+
+    const validHash = await bcrypt.compare(req.body.password, admin.password);
+    
+    if (validHash) {
+      const tokenPayLoad = {
+        sub: admin.id,
+        iat: Date.now(),
+      };
+      const token = jwt.sign(tokenPayLoad, process.env.JWT_KEY);
+      res.json({ token: token });
     } else {
-      res.json("invalid credentials");
+      adminHandler.handleAuthError(res, "Invalid credentials");
     }
   } catch (error) {
-    res.status(500).json("The server had an error");
+    adminHandler.handleServerError(res);
   }
 }
 
@@ -87,4 +115,4 @@ export default {
   createAdmin,
   deleteAdmin,
   loginAdmin,
-}
+};
